+++ src/events/configUpdateEntry.js
@@ -2,10 +2,10 @@ const { Event } = require('klasa');
 
 module.exports = class extends Event {
 
-	run(settings) {
-		if (this.client.options.sharded && settings.type === 'users') {
+	run(configs) {
+		if (this.client.options.sharded && configs.type === 'users') {
 			this.client.shard.broadcastEval(`
-				const user = this.users.get(${settings.id});
+				const user = this.users.get(${configs.id});
 				if (user) user.configs.sync();
 			`);
 		}